refactor(app): :recycle: Minor refactor of config package (broke down into separate files for better maintainability)
