+++ internal/configuration/config.go
@@ -2,8 +2,6 @@ package config
 
 import (
 	"fmt"
-	"os"
-	"regexp"
 	"strconv"
 	"strings"
 
@@ -11,6 +9,32 @@ import (
 
 	flags "github.com/richbl/go-ble-sync-cycle/internal/flags"
 )
+
+// Config represents the complete application configuration structure from the TOML config file
+type Config struct {
+	App   AppConfig   `toml:"app"`
+	BLE   BLEConfig   `toml:"ble"`
+	Speed SpeedConfig `toml:"speed"`
+	Video VideoConfig `toml:"video"`
+}
+
+// AppConfig defines application-wide settings
+type AppConfig struct {
+	LogLevel string `toml:"logging_level"`
+}
+
+// ValidationType, used for config validation, is a type that can be either an int or a float64
+type ValidationType interface {
+	int | float64
+}
+
+// validationRange is a struct used for validating config field ranges
+type validationRange struct {
+	value  any
+	min    any
+	max    any
+	errMsg error
+}
 
 // Configuration constants
 const (
@@ -29,66 +53,6 @@ const (
 	errTypeFormat = "%w: got %T"
 	errFormat     = "%w: %v"
 )
-
-// Config represents the complete application configuration structure
-type Config struct {
-	App   AppConfig   `toml:"app"`
-	BLE   BLEConfig   `toml:"ble"`
-	Speed SpeedConfig `toml:"speed"`
-	Video VideoConfig `toml:"video"`
-}
-
-// AppConfig defines application-wide settings
-type AppConfig struct {
-	LogLevel string `toml:"logging_level"`
-}
-
-// BLEConfig defines Bluetooth Low Energy settings
-type BLEConfig struct {
-	SensorBDAddr    string `toml:"sensor_bd_addr"`
-	ScanTimeoutSecs int    `toml:"scan_timeout_secs"`
-}
-
-// SpeedConfig defines speed calculation and measurement settings
-type SpeedConfig struct {
-	SmoothingWindow      int     `toml:"smoothing_window"`
-	SpeedThreshold       float64 `toml:"speed_threshold"`
-	WheelCircumferenceMM int     `toml:"wheel_circumference_mm"`
-	SpeedUnits           string  `toml:"speed_units"`
-}
-
-// VideoConfig defines video playback and display settings
-type VideoConfig struct {
-	MediaPlayer       string         `toml:"media_player"`
-	FilePath          string         `toml:"file_path"`
-	WindowScaleFactor float64        `toml:"window_scale_factor"`
-	SeekToPosition    string         `toml:"seek_to_position"`
-	UpdateIntervalSec float64        `toml:"update_interval_sec"`
-	SpeedMultiplier   float64        `toml:"speed_multiplier"`
-	OnScreenDisplay   VideoOSDConfig `toml:"OSD"`
-}
-
-// VideoOSDConfig defines on-screen display settings for video playback
-type VideoOSDConfig struct {
-	FontSize             int  `toml:"font_size"`
-	DisplayCycleSpeed    bool `toml:"display_cycle_speed"`
-	DisplayPlaybackSpeed bool `toml:"display_playback_speed"`
-	DisplayTimeRemaining bool `toml:"display_time_remaining"`
-	ShowOSD              bool // Computed field based on display settings
-}
-
-// ValidationType, used for config validation, is a type that can be either an int or a float64
-type ValidationType interface {
-	int | float64
-}
-
-// validationRange is a struct used for validating config field ranges
-type validationRange struct {
-	value  any
-	min    any
-	max    any
-	errMsg error
-}
 
 // Error messages
 var (
@@ -111,27 +75,21 @@ var (
 
 // Load loads the configuration from a TOML file using the provided flags
 func Load(configFile string) (*Config, error) {
-
-	// Note that command line flags will override values in the TOML file
 	clFlags := flags.GetFlags()
 
-	// Use the command-line config path if provided, otherwise use the default
 	if clFlags.Config != "" {
 		configFile = clFlags.Config
 	}
 
-	// Read the configuration file
 	cfg, err := readConfigFile(configFile, &Config{})
 	if err != nil {
 		return nil, err
 	}
 
-	// Validate configuration settings imported from the TOML file
 	if err := cfg.validate(); err != nil {
 		return nil, err
 	}
 
-	// Use the command-line seek position if provided (overriding any value in the TOML file)
 	if err := setSeekToPosition(cfg, clFlags); err != nil {
 		return nil, err
 	}
@@ -154,18 +112,16 @@ func readConfigFile(path string, cfg *Config) (*Config, error) {
 func setSeekToPosition(cfg *Config, clFlags flags.Flags) error {
 
 	if clFlags.Seek != "" {
-
 		if !validateTimeFormat(clFlags.Seek) {
 			return fmt.Errorf(errFormat, errInvalidSeek, clFlags.Seek)
 		}
-
 		cfg.Video.SeekToPosition = clFlags.Seek
 	}
 
 	return nil
 }
 
-// validate performs validation across all configuration sections
+// validate performs validation across all components
 func (c *Config) validate() error {
 
 	validators := []struct {
@@ -179,11 +135,9 @@ func (c *Config) validate() error {
 	}
 
 	for _, v := range validators {
-
 		if err := v.validate(); err != nil {
 			return fmt.Errorf("%s configuration error: %w", v.name, err)
 		}
-
 	}
 
 	return nil
@@ -207,153 +161,16 @@ func (ac *AppConfig) validate() error {
 	return nil
 }
 
-// validate checks BLEConfig for valid settings
-func (bc *BLEConfig) validate() error {
-
-	// Validate the scan timeout
-	if err := validateField(bc.ScanTimeoutSecs, 1, 100, errInvalidScanTimeout); err != nil {
-		return err
-	}
-
-	// Define and compile the regex for BD_ADDR
-	pattern := `^([0-9A-Fa-f]{2}(:[0-9A-Fa-f]{2}){5})$`
-	re := regexp.MustCompile(pattern)
-
-	// Check if the test string matches the pattern
-	if !re.MatchString(strings.TrimSpace(bc.SensorBDAddr)) {
-		return fmt.Errorf(errFormat, errInvalidBDAddr, bc.SensorBDAddr)
-	}
-
-	return nil
-}
-
 // validateConfigFields validates multiple fields against their min/max values
 func validateConfigFields(validations *[]validationRange) error {
 
 	for _, v := range *validations {
-
 		if err := validateField(v.value, v.min, v.max, v.errMsg); err != nil {
 			return err
 		}
-
 	}
 
 	return nil
-}
-
-// validate checks SpeedConfig for valid settings
-func (sc *SpeedConfig) validate() error {
-
-	// Validate the speed units
-	validSpeedUnits := map[string]bool{
-		SpeedUnitsKMH: true,
-		SpeedUnitsMPH: true,
-	}
-
-	if !validSpeedUnits[sc.SpeedUnits] {
-		return fmt.Errorf(errFormat, errInvalidSpeedUnits, sc.SpeedUnits)
-	}
-
-	// Create a slice of validations for range checks
-	validations := &[]validationRange{
-		{sc.SmoothingWindow, 1, 25, errSmoothingWindow},
-		{sc.SpeedThreshold, 0.0, 10.0, errSpeedThreshold},
-		{sc.WheelCircumferenceMM, 50, 3000, errWheelCircumference},
-	}
-
-	return validateConfigFields(validations)
-}
-
-// validate checks VideoConfig for valid settings
-func (vc *VideoConfig) validate() error {
-
-	// Check if the file exists
-	if _, err := os.Stat(vc.FilePath); err != nil {
-		return fmt.Errorf(errFormat, errVideoFile, err)
-	}
-
-	// Validate the media player
-	validPlayer := map[string]bool{
-		MediaPlayerMPV: true,
-		// MediaPlayerVLC: true, // temporarily disabled until vlc support is added
-	}
-
-	if !validPlayer[vc.MediaPlayer] {
-		return fmt.Errorf(errFormat, errInvalidPlayer, vc.MediaPlayer)
-	}
-
-	// Create a slice of validations for range checks
-	validations := &[]validationRange{
-		{vc.WindowScaleFactor, 0.1, 1.0, errWindowScale},
-		{vc.UpdateIntervalSec, 0.1, 3.0, errInvalidInterval},
-		{vc.SpeedMultiplier, 0.1, 1.5, errSpeedMultiplier},
-		{vc.OnScreenDisplay.FontSize, 10, 200, errFontSize},
-	}
-
-	if err := validateConfigFields(validations); err != nil {
-		return err
-	}
-
-	if !validateTimeFormat(vc.SeekToPosition) {
-		return fmt.Errorf(errFormat, errInvalidSeek, vc.SeekToPosition)
-	}
-
-	// Set ShowOSD flag based on display settings
-	vc.OnScreenDisplay.ShowOSD = vc.OnScreenDisplay.DisplayCycleSpeed ||
-		vc.OnScreenDisplay.DisplayPlaybackSpeed || vc.OnScreenDisplay.DisplayTimeRemaining
-
-	return nil
-}
-
-// validateTimeFormat checks if the provided string is valid time in MM:SS or SS format
-func validateTimeFormat(input string) bool {
-
-	input = strings.TrimSpace(input)
-
-	if strings.Contains(input, ":") {
-		return validateMMSSFormat(input)
-	}
-
-	return validateSSFormat(input)
-}
-
-// validateMMSSFormat checks if the provided string is a valid time in MM:SS format
-func validateMMSSFormat(input string) bool {
-
-	// Split the input into minutes and seconds
-	parts := strings.SplitN(input, ":", 2)
-	if len(parts) != 2 {
-		return false
-	}
-
-	minutesStr := parts[0]
-	secondsStr := parts[1]
-
-	// Validate for minutes... as long as they aren't negative
-	minutes, err := strconv.Atoi(minutesStr)
-	if err != nil || minutes < 0 {
-		return false
-	}
-
-	// Validate for seconds (0-59)
-	seconds, err := strconv.Atoi(secondsStr)
-	if err != nil || seconds < 0 || seconds > 59 {
-		return false
-	}
-
-	return true
-}
-
-// validateSSFormat checks if the provided string is a valid time in SS format
-func validateSSFormat(input string) bool {
-
-	// Validate for seconds... as long as they aren't negative
-	seconds, err := strconv.Atoi(input)
-	if err != nil || seconds < 0 {
-		return false
-	}
-
-	return true
 }
 
 // validateField checks if the provided value is within the specified range
@@ -410,4 +227,51 @@ func validateRange[T ValidationType](value, minVal, maxVal T, errMsg error) erro
 	}
 
 	return nil
+}
+
+// validateTimeFormat checks if the provided string is valid time in MM:SS or SS format
+func validateTimeFormat(input string) bool {
+
+	input = strings.TrimSpace(input)
+
+	if strings.Contains(input, ":") {
+		return validateMMSSFormat(input)
+	}
+
+	return validateSSFormat(input)
+}
+
+// validateMMSSFormat checks if the provided string is a valid time in MM:SS format
+func validateMMSSFormat(input string) bool {
+
+	parts := strings.SplitN(input, ":", 2)
+	if len(parts) != 2 {
+		return false
+	}
+
+	minutesStr := parts[0]
+	secondsStr := parts[1]
+
+	minutes, err := strconv.Atoi(minutesStr)
+	if err != nil || minutes < 0 {
+		return false
+	}
+
+	seconds, err := strconv.Atoi(secondsStr)
+	if err != nil || seconds < 0 || seconds > 59 {
+		return false
+	}
+
+	return true
+}
+
+// validateSSFormat checks if the provided string is a valid time in SS format
+func validateSSFormat(input string) bool {
+
+	seconds, err := strconv.Atoi(input)
+	if err != nil || seconds < 0 {
+		return false
+	}
+
+	return true
 }
